pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager project"
authors = ["NikolayZemelko <zemelko1989@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "task_manager" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
django-bootstrap5 = "^23.3"
flake8 = "^6.1.0"
whitenoise = "^6.5.0"
django-filter = "^23.2"
django-extensions = "^3.2.3"
rollbar = "^0.16.3"
dj-database-url = "^2.1.0"
docker = "^6.1.3"
compose = "^1.5.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
